Fix incorrect type in calls to ShapeHash
Shapehash encodings did not decode correctly because it used the wrong
lookup table when creating `ShapeHash` instances. We now use a correct
table, `json-to-shape-hash.js::types`.

```javascript
> a = {
      "firstname": "Aidan",
      "homestate": "PA"
    };
{ firstname: 'Aidan', homestate: 'PA' }

> hash = sh.toHash(a)
'0800120f0a0966697273746e616d6512020800120f0a09686f6d65737461746512020800'

> console.log(JSON.stringify(proto.decodeShapeHash(hash), null, 2))
{
  "type": "OBJECT",
  "fields": [
    {
      "key": "firstname",
      "hash": {
        "type": "OBJECT"
      }
    },
    {
      "key": "homestate",
      "hash": {
        "type": "OBJECT"
      }
    }
  ]
}
```

In this case, ["fields"]["*"]["hash"]["type"] should be "STRING".
Note that "OBJECT" happens to be the default for the field in
shapehash.proto.

Commit 4944b5d2ce081f7e4a9e86e7641c95c436e5ff31 pulled in external
protobuf files. These expect `ShapeDescriptor.type` to be a string
sourced from `json-to-shape-hash.js::types`. We accidentally used a copy
of `PrimitiveTypes` instead, prematurely converting the field into a
number. Subsequent encoding in `wire.js::_encodeShapeDescriptor` failed
lookups in `wire.js::encodePrimitiveTypes`, resulting in `undefined`
being used. This manifests as the default value of the field which,
unfortunately, is OBJECT, a valid value.

Signed-off-by: Ray Bejjani <ray.bejjani@useoptic.com>